Finished  "On the Duty of Civil Disobedience."
Moved "On the Duty of Civil Disobedience." to the completed section.
Added three books to the queue: "Daughter of the Sky: The Story of Amelia Earhart." by Paul L. Briand, Jr., "Tragedy in Dedham: The Story of the Sacco-Vanzetti Case." by Francis, "Germany Before the War." by Napoleon-Eugène Beyen Russell, and "A Century of Dishonor." by Helen Hunt Jackson.